pyproject.toml
[project]
name = "facturas"
version = "0.1.0"
description = "Proyecto sencillo que registra información de facturas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-cloud>=0.1.42",
    "llama-cloud-services>=0.6.68",
    "s3fs>=2025.9.0",
]

[dependency-groups]
demo = [
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "streamlit-authenticator>=0.4.2",
]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
where = ["./src"]
include = ["facturas"]
